+++ neo/renderer/Cinematic.cpp
@@ -1243,8 +1243,8 @@ void idCinematicLocal::readQuadInfo( byte *qData ) {
 	half = false;
 	smootheddouble = false;
 	
-	t[0] = (0 - (unsigned int)image)+(unsigned int)image+screenDelta;
-	t[1] = (0 - ((unsigned int)image + screenDelta))+(unsigned int)image;
+	t[0] = (0 - (uint64_t)image)+(uint64_t)image+screenDelta;
+	t[1] = (0 - ((uint64_t)image + screenDelta))+(uint64_t)image;
 
 	drawX = CIN_WIDTH;
 	drawY = CIN_HEIGHT;
